+++ ocrd/ocrd/task_sequence.py
@@ -36,7 +36,7 @@ class ProcessorTask():
                 tokens = tokens[3:]
             else:
                 raise Exception("Failed parsing task description '%s' with tokens remaining: '%s'" % (argstr, tokens))
-        return ProcessorTask(executable, input_file_grps, output_file_grps, parameters)
+        return cls(executable, input_file_grps, output_file_grps, parameters)
 
     def __init__(self, executable, input_file_grps, output_file_grps, parameters):
         self.executable = executable
@@ -131,7 +131,7 @@ def run_tasks(mets, log_level, page_id, task_strs, overwrite=False):
         log.info("Start processing task '%s'", task)
 
         # execute cli
-        returncode, out, err = run_cli(
+        returncode = run_cli(
             task.executable,
             mets,
             resolver,
@@ -146,7 +146,7 @@ def run_tasks(mets, log_level, page_id, task_strs, overwrite=False):
 
         # check return code
         if returncode != 0:
-            raise Exception("%s exited with non-zero return value %s. STDOUT:\n%s\nSTDERR:\n%s" % (task.executable, returncode, out, err))
+            raise Exception("%s exited with non-zero return value %s." % (task.executable, returncode))
 
         log.info("Finished processing task '%s'", task)
 
@@ -156,4 +156,4 @@ def run_tasks(mets, log_level, page_id, task_strs, overwrite=False):
         # check output file groups are in mets
         for output_file_grp in task.output_file_grps:
             if not output_file_grp in workspace.mets.file_groups:
-                raise Exception("Invalid state: expected output file group not in mets: %s\nSTDOUT:\n%s\nSTDERR:\n%s" % (output_file_grp, out, err))
+                raise Exception("Invalid state: expected output file group '%s' not in METS (despite processor success)" % output_file_grp)