+++ ocrd_models/ocrd_models/ocrd_page_generateds.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 #
-# Generated Mon Sep 21 14:00:49 2020 by generateDS.py version 2.35.20.
-# Python 3.6.9 (default, Jul 17 2020, 12:50:27)  [GCC 8.4.0]
+# Generated Fri Nov 27 15:25:12 2020 by generateDS.py version 2.35.20.
+# Python 3.6.9 (default, Oct  8 2020, 12:12:24)  [GCC 8.4.0]
 #
 # Command line options:
 #   ('-f', '')
@@ -1267,8 +1267,19 @@ class PcGtsType(GeneratedsSuper):
         Remove any empty ReadingOrder elements
         """
         ro = self.get_Page().get_ReadingOrder()
-        if ro and not ro.get_OrderedGroup() and not ro.get_UnorderedGroup():
-            self.get_Page().set_ReadingOrder(None)
+        if ro:
+            og = ro.get_OrderedGroup()
+            if og and (not og.get_RegionRefIndexed() and
+                       not og.get_OrderedGroupIndexed() and
+                       not og.get_UnorderedGroupIndexed()):
+                og = None
+            ug = ro.get_UnorderedGroup()
+            if ug and (not ug.get_RegionRef() and
+                       not ug.get_OrderedGroup() and
+                       not ug.get_UnorderedGroup()):
+                ug = None
+            if not og and not ug:
+                self.get_Page().set_ReadingOrder(None)
 # end class PcGtsType
 
 
@@ -3036,6 +3047,57 @@ class PageType(GeneratedsSuper):
                 else:
                     ret = in_reading_order + [r for r in ret if r not in in_reading_order]
         return ret
+    def invalidate_AlternativeImage(self, feature_selector=None):
+        """
+        Remove derived images from this segment (due to changed coordinates).
+    
+        If ``feature_selector`` is not none, remove only images with
+        matching ``@comments``, e.g. ``feature_selector=cropped,deskewed``.
+        """
+        existing_images = self.AlternativeImage or []
+        removed_images = []
+        if feature_selector:
+            new_images = []
+            for image in existing_images:
+                features = image.get_comments() or ''
+                if any(feature in features.split(',')
+                       for feature in feature_selector.split(',') if feature):
+                    removed_images.append(image)
+                else:
+                    new_images.append(image)
+            self.AlternativeImage = new_images
+        else:
+            removed_images = existing_images
+            self.AlternativeImage = []
+        if hasattr(self, 'id'):
+            name = self.id
+        elif hasattr(self, 'parent_object_') and hasattr(self.parent_object_, 'pcGtsId'):
+            name = self.parent_object_.pcGtsId
+        else:
+            name = ''
+        for image in removed_images:
+            self.gds_collector_.add_message('Removing AlternativeImage %s from "%s"' % (
+                image.get_comments() or '', name))
+    def set_Border(self, Border):
+        """
+        Set coordinate polygon by given object.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been cropped with a bbox
+         of the previous polygon).
+        """
+        self.invalidate_AlternativeImage(feature_selector='cropped')
+        self.Border = Border
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class PageType
 
 
@@ -3187,6 +3249,22 @@ class CoordsType(GeneratedsSuper):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_points(self, points):
+        """
+        Set coordinate polygon by given string.
+        Moreover, invalidate the parent's AlternativeImages
+        (because they will have been cropped with a bbox
+         of the previous polygon).
+        """
+        if hasattr(self, 'parent_object_'):
+            parent = self.parent_object_
+            if hasattr(parent, 'invalidate_AlternativeImage'):
+                # RegionType, TextLineType, WordType, GlyphType:
+                parent.invalidate_AlternativeImage()
+            elif hasattr(parent, 'parent_object_') and hasattr(parent.parent_object_, 'invalidate_AlternativeImage'):
+                # BorderType:
+                parent.parent_object_.invalidate_AlternativeImage(feature_selector='cropped')
+        self.points = points
 # end class CoordsType
 
 
@@ -3633,6 +3711,51 @@ class TextLineType(GeneratedsSuper):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def invalidate_AlternativeImage(self, feature_selector=None):
+        """
+        Remove derived images from this segment (due to changed coordinates).
+    
+        If ``feature_selector`` is not none, remove only images with
+        matching ``@comments``, e.g. ``feature_selector=cropped,deskewed``.
+        """
+        existing_images = self.AlternativeImage or []
+        removed_images = []
+        if feature_selector:
+            new_images = []
+            for image in existing_images:
+                features = image.get_comments() or ''
+                if any(feature in features.split(',')
+                       for feature in feature_selector.split(',') if feature):
+                    removed_images.append(image)
+                else:
+                    new_images.append(image)
+            self.AlternativeImage = new_images
+        else:
+            removed_images = existing_images
+            self.AlternativeImage = []
+        if hasattr(self, 'id'):
+            name = self.id
+        elif hasattr(self, 'parent_object_') and hasattr(self.parent_object_, 'pcGtsId'):
+            name = self.parent_object_.pcGtsId
+        else:
+            name = ''
+        for image in removed_images:
+            self.gds_collector_.add_message('Removing AlternativeImage %s from "%s"' % (
+                image.get_comments() or '', name))
+    def set_Coords(self, Coords):
+        """
+        Set coordinate polygon by given object.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been cropped with a bbox
+         of the previous polygon).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # RegionType, TextLineType, WordType, GlyphType:
+            self.invalidate_AlternativeImage()
+        elif hasattr(self, 'parent_object_') and hasattr(self.parent_object_, 'invalidate_AlternativeImage'):
+            # BorderType:
+            self.parent_object_.invalidate_AlternativeImage(feature_selector='cropped')
+        self.Coords = Coords
 # end class TextLineType
 
 
@@ -4047,6 +4170,51 @@ class WordType(GeneratedsSuper):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def invalidate_AlternativeImage(self, feature_selector=None):
+        """
+        Remove derived images from this segment (due to changed coordinates).
+    
+        If ``feature_selector`` is not none, remove only images with
+        matching ``@comments``, e.g. ``feature_selector=cropped,deskewed``.
+        """
+        existing_images = self.AlternativeImage or []
+        removed_images = []
+        if feature_selector:
+            new_images = []
+            for image in existing_images:
+                features = image.get_comments() or ''
+                if any(feature in features.split(',')
+                       for feature in feature_selector.split(',') if feature):
+                    removed_images.append(image)
+                else:
+                    new_images.append(image)
+            self.AlternativeImage = new_images
+        else:
+            removed_images = existing_images
+            self.AlternativeImage = []
+        if hasattr(self, 'id'):
+            name = self.id
+        elif hasattr(self, 'parent_object_') and hasattr(self.parent_object_, 'pcGtsId'):
+            name = self.parent_object_.pcGtsId
+        else:
+            name = ''
+        for image in removed_images:
+            self.gds_collector_.add_message('Removing AlternativeImage %s from "%s"' % (
+                image.get_comments() or '', name))
+    def set_Coords(self, Coords):
+        """
+        Set coordinate polygon by given object.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been cropped with a bbox
+         of the previous polygon).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # RegionType, TextLineType, WordType, GlyphType:
+            self.invalidate_AlternativeImage()
+        elif hasattr(self, 'parent_object_') and hasattr(self.parent_object_, 'invalidate_AlternativeImage'):
+            # BorderType:
+            self.parent_object_.invalidate_AlternativeImage(feature_selector='cropped')
+        self.Coords = Coords
 # end class WordType
 
 
@@ -4414,6 +4582,51 @@ class GlyphType(GeneratedsSuper):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def invalidate_AlternativeImage(self, feature_selector=None):
+        """
+        Remove derived images from this segment (due to changed coordinates).
+    
+        If ``feature_selector`` is not none, remove only images with
+        matching ``@comments``, e.g. ``feature_selector=cropped,deskewed``.
+        """
+        existing_images = self.AlternativeImage or []
+        removed_images = []
+        if feature_selector:
+            new_images = []
+            for image in existing_images:
+                features = image.get_comments() or ''
+                if any(feature in features.split(',')
+                       for feature in feature_selector.split(',') if feature):
+                    removed_images.append(image)
+                else:
+                    new_images.append(image)
+            self.AlternativeImage = new_images
+        else:
+            removed_images = existing_images
+            self.AlternativeImage = []
+        if hasattr(self, 'id'):
+            name = self.id
+        elif hasattr(self, 'parent_object_') and hasattr(self.parent_object_, 'pcGtsId'):
+            name = self.parent_object_.pcGtsId
+        else:
+            name = ''
+        for image in removed_images:
+            self.gds_collector_.add_message('Removing AlternativeImage %s from "%s"' % (
+                image.get_comments() or '', name))
+    def set_Coords(self, Coords):
+        """
+        Set coordinate polygon by given object.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been cropped with a bbox
+         of the previous polygon).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # RegionType, TextLineType, WordType, GlyphType:
+            self.invalidate_AlternativeImage()
+        elif hasattr(self, 'parent_object_') and hasattr(self.parent_object_, 'invalidate_AlternativeImage'):
+            # BorderType:
+            self.parent_object_.invalidate_AlternativeImage(feature_selector='cropped')
+        self.Coords = Coords
 # end class GlyphType
 
 
@@ -7029,6 +7242,20 @@ class BorderType(GeneratedsSuper):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_Coords(self, Coords):
+        """
+        Set coordinate polygon by given object.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been cropped with a bbox
+         of the previous polygon).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # RegionType, TextLineType, WordType, GlyphType:
+            self.invalidate_AlternativeImage()
+        elif hasattr(self, 'parent_object_') and hasattr(self.parent_object_, 'invalidate_AlternativeImage'):
+            # BorderType:
+            self.parent_object_.invalidate_AlternativeImage(feature_selector='cropped')
+        self.Coords = Coords
 # end class BorderType
 
 
@@ -8867,6 +9094,51 @@ class RegionType(GeneratedsSuper):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def invalidate_AlternativeImage(self, feature_selector=None):
+        """
+        Remove derived images from this segment (due to changed coordinates).
+    
+        If ``feature_selector`` is not none, remove only images with
+        matching ``@comments``, e.g. ``feature_selector=cropped,deskewed``.
+        """
+        existing_images = self.AlternativeImage or []
+        removed_images = []
+        if feature_selector:
+            new_images = []
+            for image in existing_images:
+                features = image.get_comments() or ''
+                if any(feature in features.split(',')
+                       for feature in feature_selector.split(',') if feature):
+                    removed_images.append(image)
+                else:
+                    new_images.append(image)
+            self.AlternativeImage = new_images
+        else:
+            removed_images = existing_images
+            self.AlternativeImage = []
+        if hasattr(self, 'id'):
+            name = self.id
+        elif hasattr(self, 'parent_object_') and hasattr(self.parent_object_, 'pcGtsId'):
+            name = self.parent_object_.pcGtsId
+        else:
+            name = ''
+        for image in removed_images:
+            self.gds_collector_.add_message('Removing AlternativeImage %s from "%s"' % (
+                image.get_comments() or '', name))
+    def set_Coords(self, Coords):
+        """
+        Set coordinate polygon by given object.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been cropped with a bbox
+         of the previous polygon).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # RegionType, TextLineType, WordType, GlyphType:
+            self.invalidate_AlternativeImage()
+        elif hasattr(self, 'parent_object_') and hasattr(self.parent_object_, 'invalidate_AlternativeImage'):
+            # BorderType:
+            self.parent_object_.invalidate_AlternativeImage(feature_selector='cropped')
+        self.Coords = Coords
 # end class RegionType
 
 
@@ -10722,6 +10994,17 @@ class AdvertRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class AdvertRegionType
 
 
@@ -10863,6 +11146,17 @@ class MusicRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class MusicRegionType
 
 
@@ -10976,6 +11270,17 @@ class MapRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class MapRegionType
 
 
@@ -11118,6 +11423,17 @@ class ChemRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class ChemRegionType
 
 
@@ -11260,6 +11576,17 @@ class MathsRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class MathsRegionType
 
 
@@ -11403,6 +11730,17 @@ class SeparatorRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class SeparatorRegionType
 
 
@@ -11611,6 +11949,17 @@ class ChartRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class ChartRegionType
 
 
@@ -11860,6 +12209,17 @@ class TableRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class TableRegionType
 
 
@@ -12039,6 +12399,17 @@ class GraphicRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class GraphicRegionType
 
 
@@ -12218,6 +12589,17 @@ class LineDrawingRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class LineDrawingRegionType
 
 
@@ -12410,6 +12792,17 @@ class ImageRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class ImageRegionType
 
 
@@ -12890,6 +13283,17 @@ class TextRegionType(RegionType):
         else:
             raise ValueError("Cannot hash %s" % self)
         return hash(val)
+    def set_orientation(self, orientation):
+        """
+        Set deskewing angle to given number.
+        Moreover, invalidate self's AlternativeImages
+        (because they will have been rotated and enlarged
+         with the angle of the previous value).
+        """
+        if hasattr(self, 'invalidate_AlternativeImage'):
+            # PageType, RegionType:
+            self.invalidate_AlternativeImage(feature_selector='deskewed')
+        self.orientation = orientation
 # end class TextRegionType