+++ ocrd_utils/ocrd_utils/os.py
@@ -35,7 +35,7 @@ def pushd_popd(newcwd=None, tempdir=False):
         raise Exception("pushd_popd can accept either newcwd or tempdir, not both")
     try:
         oldcwd = getcwd()
-    except FileNotFoundError as e:  # pylint: disable=unused-variable
+    except FileNotFoundError:
         # This happens when a directory is deleted before the context is exited
         oldcwd = '/tmp'
     try:
@@ -114,7 +114,7 @@ class AtomicWriterPerms(AtomicWriter):
             umask(mask)
             mode = 0o664 & ~mask
         fd = f.fileno()
-        fchmod(fd, mode)
+        chmod(fd, mode)
         return f
 
 @contextlib.contextmanager